Update miniconda.sls with working install
Using requisite onfail, passing the conda --version command with the PATH that command should reference successfully if conda is installed correctly.

Because the PATH for miniconda is directly passed to the check_conda_install command, the step ensuring that miniconda is added to PATH needs to happen every time, as that state is not ensured by check_conda_install.